+++ src/style.scss
@@ -21,6 +21,7 @@ body {
 @import 'styles/mobile';
 
 @import 'styles/buttons';
+@import 'styles/charts';
 @import 'styles/dragging';
 @import 'styles/dropdowns';
 @import 'styles/icons';
@@ -75,121 +76,4 @@ h4 {
 
 pre, code {
 	font-family: var(--font-monospace-default);
-}
-
-.usage-chart-header {
-	display: flex;
-	justify-content: space-between;
-	align-items: flex-start;
-	padding-top: 0.75rem;
-	margin-bottom: 1rem;
-
-	.usage-chart-title {
-		h3 {
-			margin-top: 0;
-			margin-bottom: 4px;
-			padding-bottom: 0;
-			border-bottom: none;
-		}
-
-		.setting-item-description {
-			margin-top: 0;
-		}
-	}
-
-	.usage-chart-controls {
-		padding-top: 4px;
-		display: flex;
-		gap: 8px;
-		select {
-			width: auto;
-		}
-	}
-}
-
-.usage-chart {
-	margin: 0;
-	padding: 1rem 0 2rem;
-	border-bottom: 1px solid var(--background-modifier-border);
-
-	.chart-line {
-		position: relative;
-		height: 80px;
-		
-		svg {
-			position: absolute;
-			top: 0;
-			left: 0;
-			right: 0;
-			bottom: 0;
-			pointer-events: none;
-			z-index: 1;
-			overflow: visible;
-			width: 100%;
-			height: 100%;
-		}
-		
-		.chart-line-path {
-			fill: none;
-			stroke: var(--chart-line-color);
-			stroke-width: 2.5;
-			vector-effect: non-scaling-stroke;
-		}
-		
-		.chart-vertical-line {
-			stroke: var(--background-modifier-border);
-			stroke-width: 1;
-			stroke-dasharray: 3;
-			vector-effect: non-scaling-stroke;
-		}
-		
-		.chart-labels {
-			position: absolute;
-			bottom: -24px;
-			left: 0;
-			right: 0;
-			display: flex;
-			justify-content: space-between;
-			color: var(--text-muted);
-			font-size: 0.72rem;
-			user-select: none;
-		}
-
-		.chart-overlay {
-			position: absolute;
-			top: 0;
-			left: 0;
-			right: 0;
-			bottom: 0;
-			z-index: 2;
-		}
-
-		.chart-tooltip {
-			position: absolute;
-			padding: 4px 8px;
-			background: var(--background-primary);
-			border: 1px solid var(--background-modifier-border);
-			box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.05);
-			border-radius: 4px;
-			pointer-events: none;
-			z-index: 100;
-			font-size: 0.72rem;
-			display: flex;
-			flex-direction: row;
-			align-items: center;
-			gap: 1.5em;
-			user-select: none;
-			
-			.tooltip-date {
-				color: var(--text-muted);
-				white-space: nowrap;
-			}
-			
-			.tooltip-count {
-				color: var(--text-accent);
-				font-weight: 600;
-				text-align: center;
-			}
-		}
-	}
 }
